requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
streamlit>=1.32
langgraph>=0.2.0
langchain-core>=0.2.0
httpx>=0.27
python-docx>=1.1.0
python-dotenv>=1.0
pydantic>=2.6
requests>=2.31
openpyxl>=3.1
